Add Invert knob to depth matte, set some defaults
Add option for inverted and not-absolute depth mattes. Also set some defaults for an Alexa Mini 4K film back @ 35mm, just so it's not got a focal length of 0 and metre scene scale as that's what I'm always using, and possibly a few others changed through use and not intent. Obviously you should reject/change those defaults if you prefer, the update is about the new knob.

Thank you so much for this kick-ass tool. It's integrated into my daily workflow and is a stellar piece of work. I've added a knob for controlling the 'positive-ness' of the generated direct depth matte, allowing it to be positive, inverted, absolute (your, and still the, default) or inverted absolute. I needed this because on a shot with heavy defocus near camera, the coc was not extending beyond the edge of the object. Since I was able to get the effect working with the standard ZDefocus, I knew there'd be a way to fix it in yours as well, and I included all of these options not because I directly needed them, but because it is more complete this way (in case of future issues of a similar nature).

I'm curious if you still have any references that you used when creating this tool... I'd be interested in learning more about the maths behind it. It's a PGBokeh-killer and for free, no less! (PGBokeh has some truly nasty pixel-level artefacts, nuke's doesn't but is nearly unusable with real-world units)